// src/templates/ts/src/builders/types.ts
export enum ParamType {
  User = 'user',
  Member = 'member',
  Channel = 'channel',
  TextChannel = 'textChannel',
  Role = 'role',
  Number = 'number',
  String = 'string',
  Boolean = 'boolean',
}

export enum FieldStyle {
  Short = 'short',
  Paragraph = 'paragraph',
}

export enum SelectType {
  User = 'user',
  Role = 'role',
  Channel = 'channel',
  Mentionable = 'mentionable',
  String = 'string',
}